Delete empty custody id lists in expiration map
PopEntryFromAwaitingSend expects that the custodyIdFlistQueue for
a given expiration in the expirationMap is nonempty.

Previously, RemoveEntryFromAwaitingSend would remove the custody
ID from the list, but would not delete empty lists from the
map.

This would lead to a segmentation fault in PopEntryFromAwaitingSend
due to a call to front() on an empty list.

This change ensures that custody ID lists in the expiration map
are non-empty if present.

RemoveEntryFromAwaitingSend is called when deleting expired
bundles. As of this commit, that's the only code-path where it is
called, which explains why this is the first time we're seeing this
issue.