Add info for ouduct initial state
Depending on the convergence layer, we may or may not know the
initial state for an outduct, i.e. whether the link is up or down.
For TCP and LTP with ping, we can quickly detect whether the link
is up or not. For LTP without ping and UDP, we do not know. Add
this info to the outuct manager and pass it to the scheduler.

The scheduler will use this info to determine whether it should
assume that a link is initially up or not.