Prevent data corruption with shared pointer
m_lastJsonAoctSharedPtr holds a shared pointer to a aoct json
string used as the backing buffer for messages previously sent
to other modules. If these sends are not complete, the std::move
here may invalidate the backing buffer before zmq is done with
the data.

I observed corrupted json messages received in ingress due to this
issue.

Fix it by using the same shared pointer idiom as in
ResendOutductCapabilities. Once the message to telem is complete,
the shared pointer will be deleted, ensuring no memory leaks while
preventing copying of the underlying data and not deleting it
before everyone is done with it.