Fix data race bug in egress
Before this change, the mutex m_mutexLinkStatusUpdate was
destroyed before the outduct manager. The outduct manager holds
the outducts, which run threads that call callbacks passed in by
egress.

These threads, via the callbacks, can try to lock this mutex
after it's destroyed, leading to an Abort in the lock code.

To fix this, destroy the outducts before any member variables.
This stops the threads and ensures the callbacks are not called
while the rest of the members are being destroyed.