Fix data race bug in OutductManager
Prior to this change, the outducts in the outduct managers could
call callbacks while the outduct manager was being destroyed. Some
member variables in the outduct manager are deleted before the
outducts are. Thus, these callbacks could possibly try to access
destroyed member variables.

We cannot simply fix this by changing the order of deletion. The
outducts are tracked in shared pointers, so all of the data
structures that hold shared pointers must be deleted before all
of the outducts are destroyed.