Split bundle update and signal gen in ctm
ctm == custody transfer manager. Need this so we can
deal with failed bundle writes to disk.

Want to be able to generate signal based on whether or
not a bundle was successfully written to disk.

So, spit the "process" method into a (1) update and (2)
generate.

Cannot just write the bundle before the update as it's
contents change. Nor can we reserve space as the size
may change.

Cannot re-generate the signal after the process call
because in the case of ACS, internal data structures
are updated. So split it in half, returning a context
struct.

This context struct holds info that the gen function
needs in order to determine what kind of signal to generate
and the contents of that signal. This info is overwritten
in the update method, so save it in this context and then
re-use it at signal generation.