Don't send status report on custody error path
When we fail to write the custody signal for a bundle, we delete
the bundle from disk, since we didn't accept custody of it. Prev.
to do this we called DeleteBundleById. But, that sends a bundle
deletion status report. We don't want that here; we just want to
remove it from the disk. So, just do that.