Fix custody double signal bug
Depending on HDTN and network settings, it's possible to queue
two of the same bundle in the egress pipeline.

How?
+ HDTN receives a with custody
+ That bundle is placed in storage
+ It is then sent to egress
+ It sits in the egress pipeline for a long time
+ The custody timer in storage expires
+ Storage sends the bundle to egress again

Note that this scenario s also possible on the larger network.
However, we'd like to prevent it from happening within HDTN.

To do that, only start the custody timer when egress sends back
an ACK indicating that the bundle was sent.

We take care to handle the case where we receive a custody signal
before we get the ack from egress (Brian indicates that this is
possible).